% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guideBed.R
\name{guideBed}
\alias{guideBed}
\title{guideBed}
\usage{
guideBed(fqBed, mgBed, outdir = NULL, sgRNA)
}
\arguments{
\item{fqBed}{the file path to fq bed file.}

\item{mgBed}{the file path to mg bed file.}

\item{outdir}{the folder of output}

\item{sgRNA}{sgRNA sequences with PAM}
}
\description{
Fetch on- and off-targets of sgRNA from bed
}
