% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intGuide.R
\name{intGuide}
\alias{intGuide}
\title{intGuide}
\usage{
intGuide(
  input,
  ref,
  sgRNA,
  bowtie2 = "bowtie2",
  samtools = "samtools",
  bedtools = "bedtools"
)
}
\arguments{
\item{input}{a folder containing raw fastq files. Not
supported file path beginning with dot, as './'!!}

\item{ref}{genome reference for bowtie2}

\item{sgRNA}{sgRNA sequences with PAM}

\item{bowtie2}{bowtie2 path}

\item{samtools}{samtools path}

\item{bedtools}{bedtools path}
}
\description{
Fetch on- and off-target from raw fastq files.
   This package integrated mergeReads, trimReads, alignBowtie2,
   sam2bam, bam2bed and guideBed, is used for a complete processing
   from raw fastq to sgRNA targets. Alternatively, you can also
   analysis your data step by step,in such mode you can tune
   lots of parameters to achieve better performance.
}
\author{
Cai Haodong
}
