% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guideBed.R
\name{preciseGRange}
\alias{preciseGRange}
\title{preciseGRange}
\usage{
preciseGRange(x, alignRange)
}
\arguments{
\item{x}{a data.frame}

\item{alignRange}{IRange object}
}
\description{
get precise ranges from relative ranges
}
